requirement.txt
torch
torchvision
pyyaml
tqdm
numpy
scipy
matplotlib
pillow